Cria a classe para abrir as configurações do telefone e seus devidos testes.
