+++ ChuckFacts/Fact/Settings/SettingsOpenable.swift
@@ -1,0 +1,14 @@
+//
+//  SettingsOpenable.swift
+//  ChuckFacts
+//
+//  Created by Calebe Emerik  | Stone on 21/03/18.
+//  Copyright © 2018 Stone Pagamentos. All rights reserved.
+//
+
+import Foundation
+
+protocol SettingsOpenable: class {
+	
+	func open()
+}
